Move the implementation of the gpuVertexFinder kernels to gpuVertexFinder.cc
